refactor: workspace service
following changes are made
1. LSPMessageDispatcher is moved to LSPMessageBroker which just manages/exposes method to manage queue.
2. LSPEventConsumer is introduced and its consumer responsibility to handle the events
3. Events are divided into notification(no response expected) and query (request expects response)
4. workspace document graph is introduced. This service maintains links between all the analysed programs and copybooks. This graph is can be referred for hover, doc definition and invalidate cache when a file event updates a copybook, thus making sure only the related docs are analysed and not all

+++ server/engine/src/main/java/org/eclipse/lsp/cobol/lsp/handlers/text/DidCloseHandler.java
@@ -20,8 +20,9 @@ import com.google.inject.Inject;
 import java.util.Objects;
 import lombok.extern.slf4j.Slf4j;
 import org.eclipse.lsp.cobol.common.copybook.CopybookService;
-import org.eclipse.lsp.cobol.lsp.AsyncAnalysisService;
 import org.eclipse.lsp.cobol.lsp.DisposableLSPStateService;
+import org.eclipse.lsp.cobol.lsp.SourceUnitGraph;
+import org.eclipse.lsp.cobol.lsp.analysis.AsyncAnalysisService;
 import org.eclipse.lsp.cobol.service.DocumentModelService;
 import org.eclipse.lsp.cobol.service.UriDecodeService;
 import org.eclipse.lsp.cobol.service.WatcherService;
@@ -39,16 +40,18 @@ public class DidCloseHandler {
   private final WatcherService watcherService;
   private final CopybookService copybookService;
   private final UriDecodeService uriDecodeService;
+  private final SourceUnitGraph sourceUnitGraph;
 
   @Inject
   public DidCloseHandler(DisposableLSPStateService disposableLSPStateService,
-                         AsyncAnalysisService asyncAnalysisService, DocumentModelService documentModelService, WatcherService watcherService, CopybookService copybookService, UriDecodeService uriDecodeService) {
+                         AsyncAnalysisService asyncAnalysisService, DocumentModelService documentModelService, WatcherService watcherService, CopybookService copybookService, SourceUnitGraph sourceUnitGraph, UriDecodeService uriDecodeService) {
     this.disposableLSPStateService = disposableLSPStateService;
     this.asyncAnalysisService = asyncAnalysisService;
     this.documentModelService = documentModelService;
     this.watcherService = watcherService;
     this.copybookService = copybookService;
     this.uriDecodeService = uriDecodeService;
+    this.sourceUnitGraph = sourceUnitGraph;
   }
 
   /**
@@ -61,8 +64,14 @@ public class DidCloseHandler {
     }
     String uri = uriDecodeService.decode(params.getTextDocument().getUri());
     LOG.info(format("Document closing invoked on URI %s", uri));
+    if (!sourceUnitGraph.isFileOpened(uri)) {
+      LOG.info(format("Ignoring document closing invoked on URI %s", uri));
+      return;
+    }
     watcherService.removeRuntimeWatchers(uri);
     documentModelService.closeDocument(uri);
+    // sourceUnitGraph.remove(uri);
+    //TODO: check if sourceUnitGraph could be used update copybook cache
     if (copybookService instanceof CopybookServiceImpl) {
       CopybookServiceImpl copybookServiceImpl = (CopybookServiceImpl) copybookService;
       copybookServiceImpl.getCopybookUsage(uri).stream()
